+++ packages/adminui-vue/components/UTableEntity/type-definitions/document.js
@@ -1,5 +1,0 @@
-/**
- * @type {UTableColumnSettings}
- */
-module.exports = {
-}