+++ packages/cs-shared/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - `formatByPattern` module with `formatNumber` and `formatDate` functions using cached **Intl** instances.
+   This module is moved from `@unitybase/ubs` to be used in apps what not require a UBS model.
 
 ### Changed